+++ game_patch/multi/server.h
@@ -10,3 +10,4 @@ void server_init();
 void server_do_frame();
 bool check_server_chat_command(const char* msg, rf::Player* sender);
 bool server_is_saving_enabled();
+void server_reliable_socket_ready(rf::Player* player);